Armor stats, armor, new rooms, scaling difficulty, scaling loot.
Armor stats, armor, new rooms, scaling difficulty, scaling loot. Some other tweaks and nice touches, too.

+++ Assets/Game/Combat/IDamageable.cs
@@ -7,7 +7,7 @@ public interface IDamageable
 {
     Team team { get; }
     int MaxHealth { get; }
-    int CurrentHealth { get; }
+    float CurrentHealth { get; }
     bool IsAlive { get; }
     void ApplyDamage(int damage);
     void Die();
