+++ Assets/Game/Decorations/Crystal/Crystal.cs
@@ -15,7 +15,7 @@ public class Crystal : MonoBehaviour, IDamageable
 
     public Team team => Team.Player;
     public int MaxHealth => 1000;
-    public int CurrentHealth { get; private set; }
+    public float CurrentHealth { get; private set; }
     public bool IsAlive => CurrentHealth > 0;
     public void ApplyDamage(int damage)
     {
